Full rewrite of the iop ordering in darktable.
This version does not use a double to order the modules but just a
simple list of modules plus the instance (multi_priority) number the
entry corresponds to.

This has make it possible to simplify a lot of part of the code and there
is no more guessing about the proper ordering. Also we do not need to
keep the exact iop-order number as before as the link between the pipe
and the module is done using the simple key module+multi-instance.

A migration path for the database is provided.

This version supports db and xmp. It remains to work on the dt styles.

As we now have full information about the module in the pipe we will be
able to provide better ordering of multi-instance when doing copy/paste
of history.

The module to change iop-order and create presets has been migrated to
this new version.

Also not that adding a new module is trivial. We just need to insert it
into the list and don't care about the iop-order value, so we can do
move module without having risk to have crashes/clashes. Again we
just handle a simple list.

+++ src/common/styles.h
@@ -40,7 +40,7 @@ typedef struct dt_style_t
 typedef struct dt_style_item_t
 {
   int num, selimg_num, enabled, multi_priority;
-  double iop_order;
+  int iop_order;
   gchar *name, *operation, *multi_name;
   int module_version, blendop_version;
   dt_iop_params_t *params;
